committing edits

+++ ft_printf.h
@@ -24,7 +24,7 @@
 int	ft_printf(const char *str, ...);
 size_t	ft_write_int(int nb);
 size_t	ft_write_uint(unsigned int nb);
-size_t	ft_size_of_num(int nb, int base);
+size_t	ft_size_of_num(size_t nb, size_t base);
 size_t	ft_write_hex(size_t nb, char *base);
 
 #endif
